+++ client/front/lib/status_runner.dart
@@ -13,7 +13,7 @@ import 'package:stomp_dart_client/stomp_dart_client.dart';
 import 'package:dropdown_button2/dropdown_button2.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'map.dart';
-import 'showerrand/re-showerrand.dart';
+import 'showerrand/re-show_errand.dart';
 final storage = FlutterSecureStorage();
 class StatusContent{//진행중인 심부름이 간략하게 담고 있는 정보들
   String contents; //메시지
